Initial Web commit + Last BackEnd adjustments before react

+++ server/src/models/graphql/index.ts
@@ -11,7 +11,7 @@ export const typeDefs = gql`
 
    type Location{
       type: String,
-      coordinates: [Int],
+      coordinates: [Float],
       is_location_exact: Boolean
    }
    type Country{
